requirements.txt
httpx
PyYAML
requests
tqdm